pyproject.toml
[project]
name = "officeagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.1",
    "asyncio>=4.0.0",
    "fastapi>=0.119.1",
    "fastmcp>=2.13.1",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langchain-google-genai>=3.0.1",
    "langchain-groq>=1.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-ollama>=1.0.0",
    "langgraph>=1.0.1",
    "livekit>=1.0.19",
    "livekit-agents[google]~=1.2",
    "livekit-plugins-deepgram>=1.2.18",
    "livekit-plugins-google>=1.2.18",
    "livekit-plugins-silero>=1.2.18",
    "ollama>=0.6.0",
    "openpyxl>=3.1.5",
    "pdfminer-six>=20250506",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
